setup/requirements.txt
cryptography
aws_cdk.core>=1.85.0,<2.0
requests